fix: Simplify string formatting in error messages in RosWorker class
