+++ autoware_system_designer/autoware_system_designer/runtime/_impl/ros2/launchers/single_node.py
@@ -12,24 +12,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Command-line builder for node_container launch type."""
+"""Command-line builder for single_node launch type."""
 
 from __future__ import annotations
 
-from typing import Mapping
+from typing import Mapping, Optional
 
-from ..common.executable import build_cmd
-from ..common.params import _ros_args
+from ..common.params import _ros_args, build_cmd, parameter_files, params_dict, remap_pairs
 
 
-def container_cmdline(spec: Mapping) -> list[str]:
+def node_cmdline(spec: Mapping, extra_param_files: Optional[list[str]] = None) -> list[str]:
     launcher = spec["launcher"]
+    inline = params_dict(spec.get("parameters", []))
+    extra_args = launcher.get("args", "")
     cmd = build_cmd(launcher)
+    if extra_args:
+        cmd += extra_args.split()
     cmd += _ros_args(
         name=spec["name"],
         namespace=spec["namespace"],
-        inline_params={},
-        param_files=[],
-        remaps=[],
+        inline_params=inline,
+        param_files=parameter_files(spec) + (extra_param_files or []),
+        remaps=remap_pairs(launcher.get("ports", [])),
     )
     return cmd
